requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
torch==2.2.1
torchvision==0.17.1
transformers==4.39.3
timm==0.9.16
facenet-pytorch==2.5.3
Pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
python-multipart==0.0.9
